Using queue instead of polling
Made mistake in last commit message, it is possible to get each
latent without having to poll or a custom diffusion pipeline through
the use of queues. Queues work across threads and get() will wait
as items come in. To make sure queue is clear for first image,
we should show it aswell instead of only through a progress bar.